Find alias hardware key rather then stopping with the software key
The DeviceInstanceID value stored in the software key appears to be
specific to newer versions of Windows so we cannot rely upon it. Instead
find the hardware key by reverse mapping via the Driver value - which is
used directly by the PnP subsystem and so should be safe to rely upon.

Signed-off-by: Paul Durrant <paul.durrant@citrix.com>