Promote xennet.sys to be a boot start driver on network boot
As a network driver, the NDIS coinstaller will normally set BootFlags
for xennet.sys.  The NDIS coinstaller does not run when performing an
offline installation.  Add BootFlags explicitly using an AddReg
directive (as is done with xenvif.sys) to allow the driver to be
available on the next boot (which may be an iSCSI boot, requiring a
functioning network).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>